+++ docs/changelog.rst
@@ -5,6 +5,8 @@ History
 -------------------
 
 * New: ``dropshadow`` :doc:`effect <effects>`.
+* New: Particle group has ``spin_drag``.
+* New: Particles can be emitted with ``angle`` and ``angle_spread``.
 
 
 1.0.0 - 2019-09-21